still wrong type

+++ users/vieting/experiments/switchboard/ctc/feat/network_helpers/specaug_configurable.py
@@ -160,10 +160,13 @@ def transform(data, network, **config):
             tf.math.floor(config["freq_mask_max_proportion"] * tf.cast(tf.shape(x)[data.feature_dim_axis], tf.float32)),
             tf.int32,
         )
-        max_time_num_seq_len = tf.math.floordiv(
-            tf.cast(tf.shape(x)[data.time_dim_axis], tf.float32),
-            max_time_num_seq_len_divisor * tf.cast(time_mask_max_size, tf.float32)
-        )
+        max_time_num_seq_len = tf.cast(
+                tf.math.floordiv(
+                    tf.cast(tf.shape(x)[data.time_dim_axis], tf.float32),
+                    max_time_num_seq_len_divisor * tf.cast(time_mask_max_size, tf.int32)
+                ),
+                tf.int32
+            )
         # check for the limits
         actual_time_mask_max_num = tf.minimum(
             tf.maximum(
